+++ bin/cli.js
@@ -27,13 +27,10 @@ program
 program
   .command('ranking')
   .description('Generate texts for dancer and country rankings.')
-  .option('-t, --top', 'Number of top ranked spots to include.')
-  .option('-w, --what', 'Ranking to use.', 'dancer')
-  .action((cmdOb, optionsArray) => {
-    const what = optionsArray[1];
-    const top = optionsArray[0];
-
-    console.log(what);
+  .option('-t, --top <int>', 'Number of top ranked spots to include.')
+  .option('-w, --what <string>', 'Ranking to use (dancer or country).', 'dancer')
+  .action((cmdOb) => {
+    const {what, top} = cmdOb;
 
     if (what === 'dancer') {
       const ranking = new DancerRanking();